Refactor: move namespace utility functions to separate file
